+++ admin/class-veritate-fact-check-crawler-admin.php
@@ -2,7 +2,7 @@
 /**
  * The admin-specific functionality of the plugin.
  *
- * @link       https://veritate.wowperations.com.br
+ * @link       https://veritatecrawler.wowperations.com.br
  * @since      0.1.0
  *
  * @package    Veritate_Fact_Check_Crawler